requirements.txt
extras==0.0.3
fuzzywuzzy==0.5.0
lettuce==0.2.20
linecache2==1.0.0
mock==1.0.1
pbr==0.11.0
python-mimeparse==0.1.4
python-subunit==1.1.0
PyYAML==3.11
six==1.9.0
sure==1.2.10
testtools==1.8.0
traceback2==1.4.0
unittest2==1.0.1
